Support Version 1.6
This makes version 1.6 the new minimum supported version. The location
of the end changed and is now limited by not just the y coordinate but
the x coordinate now as well. You need to actually hit the hook now. The
game time is now properly tracked by the game and shown at the end. It
also stays across reloading your save. This means we now use the game's
notion of the time instead of counting frames. Restarting the game in
the middle of the run and starting a new run after finishing one are
also now properly supported.